better workaround for dcpucore development
As it turns out, the previous method of specifying specific git commits in
requirements.txt didn't work -- pip will only update a package if the
package's version changed in setup.py. So, in order to force an update
every time, I added a post-compile hook that uninstalls and reinstalls
just dcpucore.

Also, I retract my statement about Heroku sucking. Besides the fact that
it was rude and entirely inappropriate, their support team was
very helpful in figuring out what the problem was, and it was only
possible to design this workaround because their buildpack is open source.